복제 함수 구현, 경고 정리
Clone() 함수 구현.
클론은 이름 그대로 복제하는 함수.
하는 일은 복사 생성자를 호출 시키는 일을 한다.

Clone()은 CEntity에서 순수 가상 함수로 정의하고,
모든 파생 클래스에 대한 예외 처리를 해주었다.
어떤 클래스는 Clone()함수를 호출하는 것을 아예 복사 생성자를 = delete;처리해서 안전하게 막아버렸고(예: 레벨)
어떤 클래스는 Clone()에 별다른 처리가 필요하지 않아서 기본 복사 생성자를 사용했고(예: 충돌체, 방향벡터),
어떤 클래스는 Clone()을 사용하기 위해서 깊은 복사가 필요했기에 복사 생성자를 직접 재정의 하고 부모를 맟추어주고...(특히 애니메이터가 까다로움) 이러한 작업이 필요했다.

일부 오류를 수정했다.
fabs()가 float을 반환하는 줄 알았는데, double을 반환해서 신기했다. abs()로 바꾸어줌.
fwprintf_s가 int를 받길래 size_t를 int로 형변환 해주었다.
이러한 자잘한 오류 들을 없애주었다.

이제 앞으로 모든 CEntity를 최상위 부모로 둔 파생 들은 CLONE 매크로를 사용해야 하고, (상황에 따라 CLONE_UNABLED를 사용하여 Clone함수를 사용하지 못하도록 막을 수 있다.) 복사 생성자를 멤버에 맞는 방식을 생각하면서 적어주어야 한다.
부모를 가리키는 포인터를 멤버로 사용할 경우 Clone()함수를 호출하는 곳에서 따로 Clone을 호출한 객체를 받아서 부모를 this로 바꿔주어야 한다.

+++ DJMAX/Client/CAnim.cpp
@@ -144,10 +144,10 @@ bool CAnim::Save(const wstring& _FilePath)
 	// 프레임 데이터
 	// 프레임 숫자
 	fwprintf_s(pFile, L"[FRAME_COUNT]\n");
-	fwprintf_s(pFile, L"%d\n\n", m_vecFrm.size());
+	fwprintf_s(pFile, L"%d\n\n", (int)m_vecFrm.size());
 
 
-	for (size_t i = 0; i < m_vecFrm.size(); ++i)
+	for (int i = 0; i < (int)m_vecFrm.size(); ++i)
 	{
 		// 프레임 정보 저장
 		fwprintf_s(pFile, L"[FRAME_NUM]\n");
@@ -214,11 +214,11 @@ bool CAnim::Load(const wstring& _FilePath)
 		}
 		else if (!wcscmp(szRead, L"[FRAME_COUNT]"))
 		{
-			size_t iFrameCount = 0;
+			int iFrameCount = 0;
 			fwscanf_s(pFile, L"%d", &iFrameCount);
 			m_vecFrm.resize(iFrameCount);
 
-			size_t iCurFrame = 0;
+			int iCurFrame = 0;
 			while (true)
 			{
 				fwscanf_s(pFile, L"%s", szRead, 256);
